+++ extensions/hibernate-reactive/deployment/src/test/java/io/quarkus/hibernate/reactive/transaction/SupportOnlyRequiredTransactionTypeTest.java
@@ -7,11 +7,11 @@ import jakarta.transaction.Transactional;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
-import io.quarkus.reactive.transaction.TransactionalInterceptorMandatory;
-import io.quarkus.reactive.transaction.TransactionalInterceptorNever;
-import io.quarkus.reactive.transaction.TransactionalInterceptorNotSupported;
-import io.quarkus.reactive.transaction.TransactionalInterceptorRequiresNew;
-import io.quarkus.reactive.transaction.TransactionalInterceptorSupports;
+import io.quarkus.reactive.transaction.runtime.TransactionalInterceptorMandatory;
+import io.quarkus.reactive.transaction.runtime.TransactionalInterceptorNever;
+import io.quarkus.reactive.transaction.runtime.TransactionalInterceptorNotSupported;
+import io.quarkus.reactive.transaction.runtime.TransactionalInterceptorRequiresNew;
+import io.quarkus.reactive.transaction.runtime.TransactionalInterceptorSupports;
 import io.quarkus.test.QuarkusUnitTest;
 import io.quarkus.test.vertx.RunOnVertxContext;
 import io.smallrye.mutiny.Uni;
