+++ extensions/reactive-transactions/deployment/src/main/java/io/quarkus/hibernate/reactive/transactions/deployment/QuarkusReactiveTransactionsProcessor.java
@@ -2,12 +2,12 @@ package io.quarkus.hibernate.reactive.transactions.deployment;
 
 import io.quarkus.arc.deployment.AdditionalBeanBuildItem;
 import io.quarkus.deployment.annotations.BuildStep;
-import io.quarkus.reactive.transaction.TransactionalInterceptorMandatory;
-import io.quarkus.reactive.transaction.TransactionalInterceptorNever;
-import io.quarkus.reactive.transaction.TransactionalInterceptorNotSupported;
-import io.quarkus.reactive.transaction.TransactionalInterceptorRequired;
-import io.quarkus.reactive.transaction.TransactionalInterceptorRequiresNew;
-import io.quarkus.reactive.transaction.TransactionalInterceptorSupports;
+import io.quarkus.reactive.transaction.runtime.TransactionalInterceptorMandatory;
+import io.quarkus.reactive.transaction.runtime.TransactionalInterceptorNever;
+import io.quarkus.reactive.transaction.runtime.TransactionalInterceptorNotSupported;
+import io.quarkus.reactive.transaction.runtime.TransactionalInterceptorRequired;
+import io.quarkus.reactive.transaction.runtime.TransactionalInterceptorRequiresNew;
+import io.quarkus.reactive.transaction.runtime.TransactionalInterceptorSupports;
 
 public class QuarkusReactiveTransactionsProcessor {
