+++ extensions/reactive-transactions/runtime/src/main/java/io/quarkus/reactive/transaction/runtime/ReactiveResource.java
@@ -1,4 +1,4 @@
-package io.quarkus.reactive.transaction;
+package io.quarkus.reactive.transaction.runtime;
 
 import io.smallrye.mutiny.Uni;
 import io.vertx.core.Context;
