+++ extensions/reactive-transactions/runtime/src/main/java/io/quarkus/reactive/transaction/runtime/TransactionalInterceptorSupports.java
@@ -1,4 +1,4 @@
-package io.quarkus.reactive.transaction;
+package io.quarkus.reactive.transaction.runtime;
 
 import jakarta.annotation.Priority;
 import jakarta.interceptor.AroundInvoke;
@@ -6,10 +6,10 @@ import jakarta.interceptor.Interceptor;
 import jakarta.interceptor.InvocationContext;
 import jakarta.transaction.Transactional;
 
-@Transactional(Transactional.TxType.NOT_SUPPORTED)
+@Transactional(Transactional.TxType.SUPPORTS)
 @Interceptor
 @Priority(Interceptor.Priority.PLATFORM_BEFORE + 300)
-public class TransactionalInterceptorNotSupported extends TransactionalInterceptorBase {
+public class TransactionalInterceptorSupports extends TransactionalInterceptorBase {
 
     @AroundInvoke
     public Object intercept(InvocationContext ic) throws Exception {
